apt install with support for both Debian 10 (Google Cloud Shell) & 11 (local)

+++ apt-install.sh
@@ -1,12 +1,19 @@
 #!/usr/bin/env bash
 set -euxo pipefail
 
+if [ $# -ne 1 ]; then
+    echo "Usage: $0 <Debian-Version> (9.0, 10, 11)"
+    echo "E.g. $0 11"
+    exit 1
+fi
+DEBIAN_VERSION=$1
+
 # https://fishshell.com =>
 # https://software.opensuse.org/download.html?project=shells%3Afish%3Arelease%3A3&package=fish
 # (because Debian 11 packages an ancient Fish v3.1.2 which is 1.5 years behind)
 sudo apt install -y curl gpg
-echo 'deb http://download.opensuse.org/repositories/shells:/fish:/release:/3/Debian_11/ /' | sudo tee /etc/apt/sources.list.d/shells:fish:release:3.list
-curl -fsSL https://download.opensuse.org/repositories/shells:fish:release:3/Debian_11/Release.key | gpg --dearmor | sudo tee /etc/apt/trusted.gpg.d/shells_fish_release_3.gpg > /dev/null
+echo "deb http://download.opensuse.org/repositories/shells:/fish:/release:/3/Debian_$DEBIAN_VERSION/ /" | sudo tee /etc/apt/sources.list.d/shells:fish:release:3.list
+curl -fsSL "https://download.opensuse.org/repositories/shells:fish:release:3/Debian_$DEBIAN_VERSION/Release.key" | gpg --dearmor | sudo tee /etc/apt/trusted.gpg.d/shells_fish_release_3.gpg > /dev/null
 
 # see install-nano.sh (which install.sh only invokes if there is no /usr/bin/nano)
 sudo apt remove -y nano
